feat(js): move babel preset to `@nrwl/js/babel` so `@nrwl/web` is not required for JS projects (e.g. React, Node)
