+++ packages/js/package.json
@@ -32,9 +32,19 @@
   "executors": "./executors.json",
   "builders": "./executors.json",
   "dependencies": {
+    "@babel/core": "^7.15.0",
+    "@babel/plugin-proposal-class-properties": "^7.14.5",
+    "@babel/plugin-proposal-decorators": "^7.14.5",
+    "@babel/plugin-transform-runtime": "^7.15.0",
+    "@babel/preset-env": "^7.15.0",
+    "@babel/preset-typescript": "^7.15.0",
+    "@babel/runtime": "^7.14.8",
     "@nrwl/devkit": "file:../devkit",
     "@nrwl/linter": "file:../linter",
     "@nrwl/workspace": "file:../workspace",
+    "babel-plugin-const-enum": "^1.0.1",
+    "babel-plugin-macros": "^2.8.0",
+    "babel-plugin-transform-typescript-metadata": "^0.3.1",
     "chalk": "^4.1.0",
     "fast-glob": "3.2.7",
     "fs-extra": "^11.1.0",
