+++ packages/web/index.ts
@@ -1,3 +1,2 @@
 export { webInitGenerator } from './src/generators/init/init';
 export { applicationGenerator } from './src/generators/application/application';
-export type { NxWebBabelPresetOptions } from './babel';
