+++ packages/web/package.json
@@ -30,13 +30,6 @@
     "migrations": "./migrations.json"
   },
   "dependencies": {
-    "@babel/core": "^7.15.0",
-    "@babel/plugin-proposal-class-properties": "^7.14.5",
-    "@babel/plugin-proposal-decorators": "^7.14.5",
-    "@babel/plugin-transform-runtime": "^7.15.0",
-    "@babel/preset-env": "^7.15.0",
-    "@babel/preset-typescript": "^7.15.0",
-    "@babel/runtime": "^7.14.8",
     "@nrwl/cypress": "file:../cypress",
     "@nrwl/devkit": "file:../devkit",
     "@nrwl/jest": "file:../jest",
@@ -44,9 +37,6 @@
     "@nrwl/linter": "file:../linter",
     "@nrwl/rollup": "file:../rollup",
     "@nrwl/workspace": "file:../workspace",
-    "babel-plugin-const-enum": "^1.0.1",
-    "babel-plugin-macros": "^2.8.0",
-    "babel-plugin-transform-typescript-metadata": "^0.3.1",
     "chalk": "^4.1.0",
     "chokidar": "^3.5.1",
     "http-server": "^14.1.0",
