+++ packages/web/src/generators/application/files/app-vite/.babelrc__tmpl__
@@ -1,5 +1,5 @@
 {
   "presets": [
-    "@nrwl/web/babel"
+    "@nrwl/js/babel"
   ]
 }
