+++ packages/workspace/src/generators/library/files/lib/__dot__babelrc__tmpl__
@@ -1,3 +1,3 @@
 {
-  "presets": [["@nrwl/web/babel", { "useBuiltIns": "usage" }]]
+  "presets": [["@nrwl/js/babel", { "useBuiltIns": "usage" }]]
 }
